==> 5_UBCF_Jupyter.ipynb <==
from surprise import KNNBasic
from surprise import Dataset
import pandas as pd
from surprise import Reader
from surprise import accuracy




sim_options = {'name': 'pearson',
               'user_based': True # compute  similarities between items
               }
algo = KNNBasic(sim_options=sim_options)


# data = pd.read_csv('Results/Train_Set_With_InjectedRatings_Length10_0.05.csv',sep=',')
data = pd.read_csv("Train_Set.csv", sep=',')

reader = Reader(rating_scale=(1, 5))


data.head()


train_data = Dataset.load_from_df(data[['X.1'  , 'X.2' , 'V1']], reader)
train_data = train_data.build_full_trainset()

del data
algo.fit(train_data)

data = pd.read_csv('Test_Set.csv',sep=',')
test_data = Dataset.load_from_df(data[['X.1'  , 'X.2' , 'V1']], reader)



test_data = test_data.build_full_trainset()
testset = test_data.build_testset()

predictions = algo.test(testset)

results_df = pd.DataFrame.from_dict(predictions)

final_results = results_df[['uid','iid','r_ui','est']].copy()


final_results.to_csv('Results/Results_Uninjected_UBCF_Python.csv',sep=',')
